Move CLI socket transport errors out of cmux monolith
