+++ CLI/CLISocketErrnoProviding.swift
@@ -1,0 +1,3 @@
+protocol CLISocketErrnoProviding {
+    var socketErrnoValue: Int32 { get }
+}
