+++ CLI/CLISocketExpectedAvailabilityError.swift
@@ -1,0 +1,1 @@
+protocol CLISocketExpectedAvailabilityError {}
