+++ CLI/CLISocketReadEOFError.swift
@@ -1,0 +1,5 @@
+struct CLISocketReadEOFError: Error, CustomStringConvertible, CLISocketExpectedAvailabilityError {
+    let message: String
+
+    var description: String { message }
+}
